New consumer script
This new consumer script subscribing all pods to a unique topic. The previous one was subscribing all pods to the same topic. 
Creating a dictionary, where the key is the topic and the value will be a deque.
This is one script that works with the producer script where one subscribe to the topics, writing to the database, while the other is listening and tried to figure out if on a specific pod.